+++ app/person-escort-record/controllers/index.js
@@ -1,0 +1,7 @@
+const FrameworkSectionController = require('./framework-section')
+const FrameworksController = require('./framework-section')
+
+module.exports = {
+  FrameworksController,
+  FrameworkSectionController,
+}
